[refactor] post 서비스 하나로 통합
