+++ src/main/java/com/sealog/backend/domain/feature/post/service/PostService.java
@@ -8,58 +8,83 @@ import com.sealog.backend.global.exception.CustomException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.util.List;
+
 /**
- * 내 게시글 서비스
+ * 게시글 서비스 인터페이스
  *
- * 인증된 사용자의 게시글 관리 비즈니스 로직
- * - 게시글 생성, 수정, 삭제, 복구
- * - 본인의 모든 상태 게시글 조회 가능
+ * 공개(Guest) 및 인증(User) 게시글 관련 비즈니스 로직
  */
-public interface PostUserService {
+public interface PostService {
+
+    // ========== Guest (공개) ========== //
+
+    /**
+     * 게시글 상세 조회 (Nickname + Slug 기반)
+     * - PUBLISHED 상태만 조회
+     *
+     * @param nickname 작성자 닉네임
+     * @param slug 조회할 게시글의 slug
+     * @return 게시글 상세 정보 (관련 게시글 포함)
+     * @throws CustomException 게시글을 찾을 수 없거나 작성자가 일치하지 않는 경우
+     */
+    PostResponse.Detail getDetail(String nickname, String slug);
+
+    /**
+     * 공개 게시글 복합 검색
+     * - PUBLISHED 상태만 검색
+     *
+     * @param condition 검색 조건 (nickname, postType, stackName, keyword)
+     * @param pageable 페이지네이션 정보
+     * @return 검색된 게시글 목록
+     */
+    Page<PostResponse.PostItems> search(PostSearchCondition condition, Pageable pageable);
+
+    /**
+     * 자동완성 검색
+     * - PUBLISHED 상태만 검색
+     *
+     * @param keyword 검색 키워드
+     * @return 검색된 게시글 목록 (최대 10개)
+     */
+    List<PostResponse.PostItems> autocomplete(String keyword);
+
+    // ========== User (인증) ========== //
 
     /**
      * 게시글 생성
-     * - 썸네일 파일 ID와 URL을 함께 전달받아 처리
-     * - 본문에서 파일 ID를 파싱하여 자동 매핑
-     * - Slug는 제목 기반 자동 생성 (중복 시 번호 추가)
      *
      * @param user 게시글 작성자
      * @param request 게시글 생성 요청 데이터
-     * @return PostResponse.Detail 생성된 게시글 상세 정보
+     * @return 생성된 게시글 상세 정보
      * @throws CustomException 제목 중복, 사용자 없음, 파일 없음 등
      */
     PostResponse.Detail create(User user, PostRequest.Create request);
 
     /**
-     * 게시글 수정
-     * - 제목 변경 시 slug 자동 재생성
-     * - 썸네일 교체 또는 제거 가능
-     * - 본문 파일 매핑 차집합 기반 자동 갱신
-     *
-     * @param userId 수정 요청 사용자 ID
-     * @param slug 수정할 게시글의 slug
-     * @param request 게시글 수정 요청 데이터
-     * @return PostResponse.Detail 수정된 게시글 상세 정보 (relatedPosts 포함)
-     * @throws CustomException 게시글 없음, 권한 없음, 제목 중복 등
-     */
-    PostResponse.Detail update(Long userId, String slug, PostRequest.Update request);
-
-    /**
      * 게시글 수정용 데이터 조회
      * - 본인 게시글은 상태 무관하게 조회 가능
-     * - relatedPosts 없이 가벼운 데이터 반환
      *
      * @param userId 조회 요청 사용자 ID
      * @param slug 조회할 게시글의 slug
-     * @return PostResponse.Edit 게시글 수정용 데이터
+     * @return 게시글 수정용 데이터
      * @throws CustomException 게시글을 찾을 수 없거나 권한이 없는 경우
      */
     PostResponse.Edit getEdit(Long userId, String slug);
 
     /**
+     * 게시글 수정
+     *
+     * @param userId 수정 요청 사용자 ID
+     * @param slug 수정할 게시글의 slug
+     * @param request 게시글 수정 요청 데이터
+     * @return 수정된 게시글 상세 정보
+     * @throws CustomException 게시글 없음, 권한 없음, 제목 중복 등
+     */
+    PostResponse.Detail update(Long userId, String slug, PostRequest.Update request);
+
+    /**
      * 게시글 삭제 (소프트 삭제)
-     * - 상태를 DELETED로 변경하고 deletedAt 기록
-     * - 실제 데이터는 DB에 유지
      *
      * @param userId 삭제 요청 사용자 ID
      * @param slug 삭제할 게시글의 slug
@@ -70,7 +95,6 @@ public interface PostUserService {
     /**
      * 게시글 복구
      * - DELETED 상태의 게시글을 PUBLISHED 상태로 복구
-     * - deletedAt 초기화
      *
      * @param userId 복구 요청 사용자 ID
      * @param slug 복구할 게시글의 slug
@@ -80,8 +104,6 @@ public interface PostUserService {
 
     /**
      * 내 게시글 검색
-     * - 본인의 게시글 조회
-     * - 게시글 타입, 스택, 키워드 등 다양한 조건으로 검색 가능
      *
      * @param userId 조회 요청 사용자 ID
      * @param condition 검색 조건 (postType, stackName, keyword)
@@ -92,12 +114,10 @@ public interface PostUserService {
 
     /**
      * 삭제된 게시글 조회
-     * - DELETED 상태의 게시글만 조회
-     * - 복구 기능을 위해 제공
      *
      * @param userId 조회 요청 사용자 ID
      * @param pageable 페이지네이션 정보
-     * @return Page<PostResponse.PostItems> 삭제된 게시글 목록
+     * @return 삭제된 게시글 목록
      */
     Page<PostResponse.PostItems> getDeleted(Long userId, Pageable pageable);
 }
